cleanup documentation
The synopsis and options sections are displayed when you provide
a bad option.  The most useful information in this case is a simple
usage and detailed description of the options.

Keep the old synopsis as an EXAMPLE section.  reword some of the
main body, and include links to other modules where appropriate.